Rename fra avtalt-form til ActaltForm
