+++ src/moduler/aktivitet/visning/avtalt-container/forhaandsorienterings-melding.tsx
@@ -6,7 +6,7 @@ import Select from '../../../../felles-komponenter/skjema/input/select';
 import Textarea from '../../../../felles-komponenter/skjema/input/Textarea';
 import InternLenke from '../../../../felles-komponenter/utils/InternLenke';
 import VisibleIfDiv from '../../../../felles-komponenter/utils/visible-if-div';
-import { IKKE_SEND_FORHANDSORIENTERING, SEND_FORHANDSORIENTERING, SEND_PARAGRAF_11_9 } from './avtalt-form';
+import { IKKE_SEND_FORHANDSORIENTERING, SEND_FORHANDSORIENTERING, SEND_PARAGRAF_11_9 } from './AvtaltForm';
 
 function InfoHeader() {
     return (
